+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The changelog for `SuperwallKit`. Also see the [releases](https://github.com/sup
 ### Fixes
 
 - Fixes issue where holdouts were still matching even if the limit set for their corresponding rules were exceeded.
+- Fixes potential crash if the free trial notification delay was set to zero seconds.
 
 ## 3.4.4